+++ ttnn/cpp/ttnn/operations/data_movement/data_movement_pybind.hpp
@@ -10,6 +10,7 @@
 #include "ttnn/cpp/pybind11/decorators.hpp"
 #include "ttnn/operations/data_movement.hpp"
 #include "ttnn/operations/data_movement/pad/pad_pybind.hpp"
+#include "ttnn/operations/data_movement/concat/concat_pybind.hpp"
 
 namespace py = pybind11;
 
@@ -39,38 +40,6 @@ Example:
         &permute,
         py::arg("input_tensor"),
         py::arg("order"),
-        doc);
-}
-
-void bind_concat(py::module& module) {
-    const auto doc = R"doc(
-Concats :attr:`tensors` in the given :attr:`dim`.
-
-Args:
-    * :attr:`tensors`: the tensors to be concatenated.
-    * :attr:`dim`: the concatenating dimension.
-
-Keyword Args:
-    * :attr:`memory_config`: the memory configuration to use for the operation
-
-Example:
-
-    >>> tensor = ttnn.concat(ttnn.from_torch(torch.zeros((1, 1, 64, 32), ttnn.from_torch(torch.zeros((1, 1, 64, 32), dim=3)), device)
-
-    >>> tensor1 = ttnn.from_torch(torch.zeros((1, 1, 64, 32), dtype=torch.bfloat16), device=device)
-    >>> tensor2 = ttnn.from_torch(torch.zeros((1, 1, 64, 32), dtype=torch.bfloat16), device=device)
-    >>> output = ttnn.concat([tensor1, tensor2], dim=4)
-    >>> print(output.shape)
-    [1, 1, 32, 64]
-
-    )doc";
-    module.def(
-        "concat",
-        &concat,
-        py::arg("input_tensor"),
-        py::arg("dim") = 0,
-        py::kw_only(),
-        py::arg("memory_config") = std::nullopt,
         doc);
 }
 
@@ -173,7 +142,7 @@ Example:
 
 void py_module(py::module& module) {
     bind_permute(module);
-    bind_concat(module);
+    detail::bind_concat(module);
     bind_upsample(module);
     detail::bind_pad(module);
     bind_repeat(module);
